feat: Improve project structure and documentation
This commit improves the project structure by refactoring the main web application into separate files for UI, utilities, and styling. This enhances code organization and maintainability.

Key changes:
- Created `ui.py` for the Streamlit user interface.
- Created `utils.py` for data loading and model prediction functions.
- Added `style.css` for custom styling of the web app.
- Introduced `crop_info.json` to store crop-related information, making it easier to manage and extend.
- Added a new Jupyter Notebook (`KrushiAI.ipynb`) with detailed data analysis, model training, and evaluation. This notebook provides a comprehensive overview of the project's development process.
- Updated the `README.md` file with a more detailed explanation of the project, including the new file structure, updated model accuracy, and embedded charts for better visualization.
- Removed the old, unstructured Jupyter Notebook (`KrushiAI_Crop_Recommendation.ipynb`) and replaced it with the improved version.
