+++ crop_info.json
@@ -1,0 +1,24 @@
+{
+    "rice": "Rice thrives in warm, humid conditions with abundant water. Ideal for lowland areas with good irrigation.",
+    "maize": "Maize (corn) grows well in well-drained soils with moderate rainfall and warm temperatures.",
+    "chickpea": "Chickpeas prefer cool, dry conditions and can tolerate drought. They enrich soil with nitrogen.",
+    "kidneybeans": "Kidney beans need warm temperatures and moderate rainfall. They prefer well-drained, fertile soil.",
+    "pigeonpeas": "Pigeon peas are drought-resistant and grow well in semi-arid regions with minimal rainfall.",
+    "mothbeans": "Moth beans are extremely drought-tolerant and thrive in hot, dry conditions with minimal water.",
+    "mungbean": "Mung beans prefer warm temperatures and moderate rainfall. They have a short growing season.",
+    "blackgram": "Black gram thrives in warm, humid conditions and can tolerate some drought.",
+    "lentil": "Lentils prefer cool growing conditions and moderate rainfall. They're adaptable to various soil types.",
+    "pomegranate": "Pomegranates thrive in hot, dry climates and are drought-tolerant once established.",
+    "banana": "Bananas need consistent warmth, high humidity, and abundant water. They're sensitive to frost.",
+    "mango": "Mangoes require tropical conditions with a distinct dry season for flowering. They're frost-sensitive.",
+    "grapes": "Grapes grow best in temperate climates with warm, dry summers and mild winters.",
+    "watermelon": "Watermelons need hot temperatures, plenty of sunlight, and moderate water during growth.",
+    "muskmelon": "Muskmelons require warm temperatures, full sun, and moderate, consistent moisture.",
+    "apple": "Apples need a cold winter period for dormancy and moderate summers. They prefer well-drained soil.",
+    "orange": "Oranges thrive in subtropical climates with mild winters and warm summers.",
+    "papaya": "Papayas need consistent warmth and moisture. They're very frost-sensitive.",
+    "coconut": "Coconuts require tropical conditions with high humidity, warm temperatures, and regular rainfall.",
+    "cotton": "Cotton thrives in warm climates with long growing seasons and moderate rainfall.",
+    "jute": "Jute needs warm, humid conditions with high rainfall during the growing season.",
+    "coffee": "Coffee grows best in tropical highlands with moderate temperatures and regular rainfall."
+}
